Returns git at index

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
  * Return: bit at index
  */
 
-int get_bit(unsigned int n, unsigned int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned int temp;
